requirements.txt
fastapi==0.110.0
starlette==0.36.3
uvicorn[standard]==0.29.0
httpx==0.28.1
pydantic==1.10.24
jinja2==3.1.6
SQLAlchemy==2.0.44
python-dotenv==1.2.1
openai==1.12.0
python-dateutil==2.9.0.post0
spacy==3.8.11
